--- src/app/all-classes.spec.ts
@@ -1,7 +1,0 @@
-import { AllClasses } from './all-classes';
-
-describe('AllClasses', () => {
-  it('should create an instance', () => {
-    expect(new AllClasses()).toBeTruthy();
-  });
-});
